Improve CLI tests.

+++ vsentry-cli/test.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python2
+
 import subprocess
 import argparse
 
@@ -114,13 +116,15 @@ def check_actions():
 	create_action('action1', 'drop', 'syslog', 'none', 'none')
 	create_action('action1', 'allow', 'none', 'none', 'none')
 	create_action('action1', 'allow', 'file', 'drop', 'syslog')
+	create_action('action2', 'allow', 'file', 'drop', 'syslog')
 
 def delete_rule(rule_type, section, rule_number):
 	delete_cmd = cli_cmd + ' delete ' + rule_type + ' ' + section + ' rule_number ' + str(rule_number)
 	try:
 		run_cmd(delete_cmd)
 	except subprocess.CalledProcessError, e:
-		print 'info: failed delete rule ' + str(rule_number)
+		if is_verbose:
+			print 'info: failed delete rule ' + str(rule_number)
 
 def add_rule_field(cmd, is_group, field, value):
 	new_cmd = cmd + ' ' + field
@@ -182,9 +186,10 @@ def is_valid_can_rule(reply, is_mid_group, mid, can_dir, is_if_group, interface,
 		
 	return True
 	
-def check_can_rule_add(rule_type, rule_number, is_mid_group, mid, can_dir, is_if_group, interface, is_program_group, program, is_user_group, user, action):
+def check_can_rule_add(rule_type, rule_number, is_mid_group, mid, can_dir, is_if_group, interface, is_program_group, program, is_user_group, user, action, is_success, is_delete):
 	global errs
-	delete_rule(rule_type, 'can', rule_number)
+	if is_delete:
+		delete_rule(rule_type, 'can', rule_number)
 	update_cmd = cli_cmd + ' update ' + rule_type + ' can ' + ' rule_number ' + str(rule_number)
 	update_cmd = add_rule_field(update_cmd, is_mid_group, 'mid', mid)
 	update_cmd += ' dir ' + can_dir
@@ -193,14 +198,21 @@ def check_can_rule_add(rule_type, rule_number, is_mid_group, mid, can_dir, is_if
 	update_cmd = add_rule_field(update_cmd, is_user_group, 'user', user)
 	update_cmd += ' action ' + action
 	if is_verbose:
-		print update_cmd
+		print 'check CAN add rule: ', update_cmd
 	try:
 		run_cmd(update_cmd)
 	except subprocess.CalledProcessError, e:
-		print 'Failed add can rule : ' + str(rule_number)
+		if (is_success):
+			print 'Failed add can rule : ' + str(rule_number)
+			errs += 1
+		return
+	if not is_success:
+		print 'ERROR - invalud rule success, rule:', rule_number
 		errs += 1
 		return
 	reply = get_rule(rule_type, 'can', rule_number)
+	if is_verbose:
+		print 'reply: ', reply
 	if (not is_valid_can_rule(reply, is_mid_group, mid, can_dir, is_if_group, interface, is_program_group, program, is_user_group, user, action)):
 		print 'ERROR in CAN rule number:', str(rule_number)
 		errs += 1
@@ -239,7 +251,7 @@ def is_valid_file_rule(reply, is_file_group, filename, perm, is_program_group, p
 		return False
 	return True
 
-def check_file_rule_add(rule_type, rule_number, is_file_group, filename, perm, is_program_group, program, is_user_group, user, action):
+def check_file_rule_add(rule_type, rule_number, is_file_group, filename, perm, is_program_group, program, is_user_group, user, action, is_success):
 	global errs
 	delete_rule(rule_type, 'file', rule_number)
 	update_cmd = cli_cmd + ' update ' + rule_type + ' file ' + ' rule_number ' + str(rule_number)
@@ -251,7 +263,13 @@ def check_file_rule_add(rule_type, rule_number, is_file_group, filename, perm, i
 	try:
 		run_cmd(update_cmd)
 	except subprocess.CalledProcessError, e:
-		print 'ERROR Failed add file rule : ' + str(rule_number)
+		if is_success:
+			print 'ERROR Failed add file rule : ' + str(rule_number)
+			errs += 1
+		return
+	if not is_success:
+		print 'ERROR success of invalid rule number:', rule_number
+		print update_cmd
 		errs += 1
 		return
 	reply = get_rule(rule_type, 'file', rule_number)
@@ -261,17 +279,28 @@ def check_file_rule_add(rule_type, rule_number, is_file_group, filename, perm, i
 
 def check_can_rules():
 	print '---------------- Check CAN Rules'
-	check_can_rule_add('rule', 10, True, 'mid_group1', 'in', True, 'caninf_group1', True, 'program_group1', True, 'user_group1', 'action1')
-	check_can_rule_add('rule', 10, False, '123', 'out', True, 'caninf_group1', True, 'program_group1', True, 'user_group1', 'action1')
-	check_can_rule_add('rule', 10, False, '123', 'both', False, 'vcan0', True, 'program_group1', True, 'user_group1', 'action1')
-	check_can_rule_add('rule', 10, False, '123', 'both', False, 'vcan0', False, '/bin/cat', True, 'user_group1', 'action1')
-	check_can_rule_add('rule', 10, False, '123', 'both', False, 'vcan0', False, '/bin/cat', False, 'root', 'action1')
+	check_can_rule_add('rule', 11, True, 'mid_group6', 'in', True, 'caninf_group1', True, 'program_group1', True, 'user_group1', 'action1', False, True)
+	check_can_rule_add('rule', 11, True, 'mid_group1', 'both', True, 'caninf_grou', True, 'program_group1', True, 'user_group1', 'action1', False, True)
+	check_can_rule_add('rule', 10, True, 'mid_group1', 'in', True, 'caninf_group1', True, 'program_group1', True, 'user_group1', 'action1', True, True)
+	check_can_rule_add('rule', 10, False, '123', 'out', True, 'caninf_group1', True, 'program_group1', True, 'user_group1', 'action1', True, True)
+	check_can_rule_add('rule', 10, False, '123', 'both', False, 'vcan0', True, 'program_group1', True, 'user_group1', 'action1', True, True)
+	check_can_rule_add('rule', 10, False, '123', 'both', False, 'vcan0', False, '/bin/cat', True, 'user_group1', 'action1', True, True)
+	check_can_rule_add('rule', 10, False, '123', 'both', False, 'vcan0', False, '/bin/cat', False, 'root', 'action1', True, True)
+	check_can_rule_add('rule', 10, False, '124', 'both', False, 'vcan0', False, '/bin/cat', False, 'root', 'action1', True, False)
+	check_can_rule_add('rule', 10, False, '124', 'in', False, 'vcan0', False, '/bin/cat', False, 'root', 'action1', True, False)
+	check_can_rule_add('rule', 10, False, 'any', 'in', False, 'vcan0', False, '/bin/cat', False, 'root', 'action1', True, False)
+	check_can_rule_add('rule', 10, False, 'any', 'in', False, 'vcan1', False, '/bin/cat', False, 'root', 'action1', True, False)
+	check_can_rule_add('rule', 10, False, 'any', 'in', False, 'vcan1', False, '/bin/echo', False, 'root', 'action1', True, False)
+	check_can_rule_add('rule', 10, False, 'any', 'in', False, 'vcan1', False, '/bin/echo', False, '*', 'action1', True, False)
+	check_can_rule_add('rule', 10, False, 'any', 'in', False, 'vcan1', False, '/bin/echo', False, '*', 'action2', True, False)
 
 def check_file_rules():
 	print '---------------- Check FILE Rules'
-	check_file_rule_add('rule', 10, True, 'file_group1', 'r', True, 'program_group1', True, 'user_group1', 'action1')
-	check_file_rule_add('rule', 10, False, '/work/file1.txt' , 'rw', True, 'program_group1', True, 'user_group1', 'action1')
-	check_file_rule_add('rule', 10, False, '/work/file1.txt' , 'rw', False, '/bin/cat', False, 'root', 'action1')
+	check_file_rule_add('rule', 11, True, 'invalid_group' , 'rw', False, '/bin/cat', False, 'root', 'action1', False)
+	check_file_rule_add('rule', 11, False, '/work/file1.txt' , 'rw', True, 'invalid_group', False, 'root', 'action1', False)
+	check_file_rule_add('rule', 10, True, 'file_group1', 'r', True, 'program_group1', True, 'user_group1', 'action1', True)
+	check_file_rule_add('rule', 10, False, '/work/file1.txt' , 'rw', True, 'program_group1', True, 'user_group1', 'action1', True)
+	check_file_rule_add('rule', 10, False, '/work/file1.txt' , 'rw', False, '/bin/cat', False, 'root', 'action1', True)
 
 def is_valid_ip_rule(reply, is_src_addr_group, src_addr, is_dst_addr_group, dst_addr, is_proto_group, proto, is_src_port_group, src_port, is_dst_port_group, dst_port, is_program_group, program, is_user_group, user, action):
 	rule_src_addr = reply[1]
@@ -312,7 +341,7 @@ def is_valid_ip_rule(reply, is_src_addr_group, src_addr, is_dst_addr_group, dst_
 		return False
 	return True
 
-def check_ip_rule_add(rule_type, rule_number, is_src_addr_group, src_addr, is_dst_addr_group, dst_addr, is_proto_group, proto, is_src_port_group, src_port, is_dst_port_group, dst_port, is_program_group, program, is_user_group, user, action):
+def check_ip_rule_add(rule_type, rule_number, is_src_addr_group, src_addr, is_dst_addr_group, dst_addr, is_proto_group, proto, is_src_port_group, src_port, is_dst_port_group, dst_port, is_program_group, program, is_user_group, user, action, is_success):
 	delete_rule(rule_type, 'ip', rule_number)
 	update_cmd = cli_cmd + ' update ' + rule_type + ' ip ' + ' rule_number ' + str(rule_number)
 	update_cmd = add_rule_field(update_cmd, is_src_addr_group, 'src_addr', src_addr)
@@ -326,7 +355,12 @@ def check_ip_rule_add(rule_type, rule_number, is_src_addr_group, src_addr, is_ds
 	try:
 		run_cmd(update_cmd)
 	except subprocess.CalledProcessError, e:
-		print 'ERROR Failed add ip rule : ' + str(rule_number)
+		if is_success:
+			print 'ERROR Failed add ip rule : ' + str(rule_number)
+			errs += 1
+		return
+	if not is_success:
+		print 'ERROR invalid rule sucees, rule id:' + str(rule_number)
 		errs += 1
 		return
 	reply = get_rule(rule_type, 'ip', rule_number)
@@ -336,9 +370,9 @@ def check_ip_rule_add(rule_type, rule_number, is_src_addr_group, src_addr, is_ds
 	
 def check_ip_rules():
 	print '---------------- Check IP Rules'
-	check_ip_rule_add('rule', '10', True, 'addr_group1', True, 'addr_group2', True, 'proto_group1', True, 'port_group1', True, 'port_group2', True, 'program_group1', True, 'user_group1', 'action1')
-	check_ip_rule_add('rule', '10', False, '6.5.4.3/24', True, 'addr_group2', True, 'proto_group1', True, 'port_group1', True, 'port_group2', True, 'program_group1', True, 'user_group1', 'action1')
-	check_ip_rule_add('rule', '10', False, '6.5.4.3/24', False, '1.1.1.1/32', True, 'proto_group1', True, 'port_group1', True, 'port_group2', True, 'program_group1', True, 'user_group1', 'action1')
+	check_ip_rule_add('rule', '10', True, 'addr_group1', True, 'addr_group2', True, 'proto_group1', True, 'port_group1', True, 'port_group2', True, 'program_group1', True, 'user_group1', 'action1', True)
+	check_ip_rule_add('rule', '10', False, '6.5.4.3/24', True, 'addr_group2', True, 'proto_group1', True, 'port_group1', True, 'port_group2', True, 'program_group1', True, 'user_group1', 'action1', True)
+	check_ip_rule_add('rule', '10', False, '6.5.4.3/24', False, '1.1.1.1/32', True, 'proto_group1', True, 'port_group1', True, 'port_group2', True, 'program_group1', True, 'user_group1', 'action1', True)
 
 def check_rules():
 	check_can_rules()
